backend/requirements.txt
fastapi
uvicorn
pymupdf
python-docx
google-generativeai
python-multipart
python-dotenv
